+++ app/hubuki/.task/checksum/fmt
@@ -1,0 +1,1 @@
+5f153d3d620c7c1c8877982df37175f9
