+++ app/hubuki/app/main.go
@@ -4,23 +4,16 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+
+	"github.com/shunsock/hubuki/internal/server"
 )
 
 func main() {
-	http.HandleFunc("/", HomeHandler)
-	http.HandleFunc("/health", HealthHandler)
+	http.HandleFunc("/", server.HomeHandler)
+	http.HandleFunc("/health", server.HealthHandler)
 
 	fmt.Println("Starting server on :8080")
 	if err := http.ListenAndServe(":8080", nil); err != nil {
 		log.Fatal("Server failed to start:", err)
 	}
-}
-
-func HomeHandler(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, "Hello, Hubuki! 🎋\n")
-}
-
-func HealthHandler(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-	fmt.Fprintf(w, "OK")
 }
