+++ app/hubuki/internal/server/server.go
@@ -1,0 +1,15 @@
+package server
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func HomeHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintf(w, "Hello, Hubuki! 🎋\n")
+}
+
+func HealthHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+	fmt.Fprintf(w, "OK")
+}
